pyproject.toml
[tool.poetry]
name = "tga to png"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyinstaller = {url = "https://github.com/rokm/pyinstaller/archive/refs/heads/python-3.10.zip"}
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
